upgrade view and schema classes
